Renames package util.android to androidUtil
And moves GnomyFragmentFactory and GnomySpinnerAdapter to
that package.
