+++ app/src/androidTest/java/io/github/alansanchezp/gnomy/account/ArchivedAccountsDialogFragmentInstrumentedTest.java
@@ -13,7 +13,7 @@ import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import io.github.alansanchezp.gnomy.R;
 import io.github.alansanchezp.gnomy.database.account.Account;
-import io.github.alansanchezp.gnomy.ui.GnomyFragmentFactory;
+import io.github.alansanchezp.gnomy.androidUtil.GnomyFragmentFactory;
 import io.github.alansanchezp.gnomy.ui.account.ArchivedAccountsDialogFragment;
 
 import static androidx.fragment.app.testing.FragmentScenario.launchInContainer;
