+++ app/src/main/java/io/github/alansanchezp/gnomy/androidUtil/SingleClickViewHolder.java
@@ -1,4 +1,4 @@
-package io.github.alansanchezp.gnomy.util.android;
+package io.github.alansanchezp.gnomy.androidUtil;
 
 import android.app.Activity;
 import android.os.SystemClock;
