+++ app/src/main/java/io/github/alansanchezp/gnomy/ui/transaction/TransactionFiltersDialogFragment.java
@@ -41,17 +41,17 @@ import io.github.alansanchezp.gnomy.database.category.Category;
 import io.github.alansanchezp.gnomy.database.transaction.MoneyTransaction;
 import io.github.alansanchezp.gnomy.database.transaction.MoneyTransactionFilters;
 import io.github.alansanchezp.gnomy.databinding.DialogTransactionFiltersBinding;
-import io.github.alansanchezp.gnomy.ui.GnomyFragmentFactory;
-import io.github.alansanchezp.gnomy.ui.GnomySpinnerAdapter;
+import io.github.alansanchezp.gnomy.androidUtil.GnomyFragmentFactory;
+import io.github.alansanchezp.gnomy.androidUtil.GnomySpinnerAdapter;
 import io.github.alansanchezp.gnomy.util.BigDecimalUtil;
 import io.github.alansanchezp.gnomy.util.ColorUtil;
 import io.github.alansanchezp.gnomy.util.DateUtil;
-import io.github.alansanchezp.gnomy.util.android.InputFilterMinMax;
-import io.github.alansanchezp.gnomy.util.android.SingleClickViewHolder;
-import io.github.alansanchezp.gnomy.util.android.ViewTintingUtil;
+import io.github.alansanchezp.gnomy.androidUtil.InputFilterMinMax;
+import io.github.alansanchezp.gnomy.androidUtil.SingleClickViewHolder;
+import io.github.alansanchezp.gnomy.androidUtil.ViewTintingUtil;
 
 import static io.github.alansanchezp.gnomy.util.ISpinnerItem.getItemIndexById;
-import static io.github.alansanchezp.gnomy.util.android.SimpleTextWatcherWrapper.onlyOnTextChanged;
+import static io.github.alansanchezp.gnomy.androidUtil.SimpleTextWatcherWrapper.onlyOnTextChanged;
 
 public class TransactionFiltersDialogFragment
         extends DialogFragment
